donations/donation-received/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-secret-manager==2.*
google-auth==2.*
google-api-python-client==2.*
requests==2.*
